try add water(with bugs)

+++ DevilApartment/src/pixel/common_dy.rs
@@ -1,2 +1,2 @@
 pub const DY_LUT_LEN: usize = 16;
-pub const DY_LUT: [usize; (DY_LUT_LEN as usize)] = [1, 1, 2, 2, 3, 3, 4, 4, 5, 5, 6, 6, 7, 7, 8, 8];
+pub const DY_LUT: [usize; DY_LUT_LEN] = [1, 1, 2, 2, 3, 3, 4, 4, 5, 5, 6, 6, 7, 7, 8, 8];
